edit note in separate module
