+++ app/src/main/java/com/stslex93/notes/NoteApplication.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import com.stslex93.notes.core.database.coreDatabaseModule
 import com.stslex93.notes.core.notes.di.coreNotesModule
 import com.stslex93.notes.di.appModule
+import com.stslex93.notes.feature.edit.di.featureEditModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -22,7 +23,8 @@ class NoteApplication : Application() {
             modules(
                 appModule,
                 coreDatabaseModule,
-                coreNotesModule
+                coreNotesModule,
+                featureEditModule
             )
         }
     }
