+++ app/src/main/res/navigation/nav_graph.xml
@@ -16,7 +16,7 @@
 
     <fragment
         android:id="@+id/nav_edit"
-        android:name="com.stslex93.notes.ui.edit.EditFragment">
+        android:name="com.stslex93.notes.feature.edit.ui.EditFragment">
         <argument
             android:name="id"
             app:argType="integer" />
