+++ feature/edit/src/main/java/com/stslex93/notes/feature/edit/domain/interactor/NoteEditInteractor.kt
@@ -1,6 +1,6 @@
-package com.stslex93.notes.domain.interactor.interf
+package com.stslex93.notes.feature.edit.domain.interactor
 
-import com.stslex93.notes.domain.model.NoteDomainModel
+import com.stslex93.notes.feature.edit.domain.model.NoteDomainModel
 import kotlinx.coroutines.flow.Flow
 
 interface NoteEditInteractor {
