+++ feature/edit/src/main/java/com/stslex93/notes/feature/edit/domain/model/NoteDomainModel.kt
@@ -1,0 +1,8 @@
+package com.stslex93.notes.feature.edit.domain.model
+
+data class NoteDomainModel(
+    val id: Int,
+    val title: String,
+    val content: String,
+    val timestamp: Long,
+)
