+++ feature/edit/src/main/java/com/stslex93/notes/feature/edit/ui/EditNoteViewModel.kt
@@ -1,7 +1,7 @@
-package com.stslex93.notes.ui.edit
+package com.stslex93.notes.feature.edit.ui
 
 import com.stslex93.notes.core.ui.base.BaseViewModel
-import com.stslex93.notes.ui.edit.store.EditStore
+import com.stslex93.notes.feature.edit.ui.store.EditStore
 
 class EditNoteViewModel(
     store: EditStore
