+++ feature/edit/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="title_hint">Title</string>
+    <string name="content_hint">Content</string>
+    <string name="label_edit">Edit</string>
+</resources>
